Final Change (Completed)
Last Changes for Project. (Completed)
Added connection to database and insertion of  usernames ,scores and
date of the inserted user to the database. Created admin that can see
the score to every question and make the quiz as many times as he wants
Leaderboards gets data from database with  jstl .Also there is a page
that redirects to various pages for the user to see the mistakes
depending on the mistakes the user did.